RestaurantsList and RestaurantsFilters connected to DB.
