+++ src/app/components/restaurants-list/restaurants-list.component.ts
@@ -10,7 +10,7 @@ import { FilterService } from '../../services/filter.service';
   styleUrls: ['./restaurants-list.component.scss']
 })
 export class RestaurantsListComponent implements OnInit {
-  restaurants: ICafe[];
+  restaurants: ICafe[] = [];
   categoryFilter: String[] = [];
   cuisineFilter: String[] = [];
   featureFilter: String[] = [];
@@ -36,7 +36,13 @@ export class RestaurantsListComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.restaurants = this.getCafesService.getAllCafes();
+    this.getCafesService.getAllCafes()
+      .subscribe(
+        (restaurants: ICafe[]) => {
+          this.restaurants = restaurants;
+        },
+        (error) => console.log(error)
+      );
     this.filterService.currentCafes.subscribe(data => this.restaurants = data);
   }
